third_party/WebKit/LayoutTests/fast/css/parsing-object-fit-expected.txt

This tests checks that all of the input values for object-fit parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testComputedStyle(";") is "fill"
PASS test("object-fit: inherit;") is "inherit"
PASS test("object-fit: initial;") is "initial"
PASS test("object-fit: fill;") is "fill"
PASS test("object-fit: contain;") is "contain"
PASS test("object-fit: cover;") is "cover"
PASS test("object-fit: none;") is "none"
PASS test("object-fit: scale-down;") is "scale-down"
PASS test("object-fit: fill contain;") is ""
PASS test("object-fit: bananas;") is ""
PASS test("object-fit: 23px;") is ""
PASS test("object-fit: 20%;") is ""
PASS successfullyParsed is true

TEST COMPLETE
